Use uint8_t in framebuffer

+++ include/frameBuffer.hpp
@@ -13,13 +13,13 @@ public:
     uint32_t width();
     uint32_t height();
     void     resize(uint32_t width, uint32_t height);
-    float*   pixelArray();
+    uint8_t*   pixelArray();
     void     display();
 
 private:
-    uint32_t           _width;
-    uint32_t           _height;
-    std::vector<float> _pixels;
+    uint32_t             _width;
+    uint32_t             _height;
+    std::vector<uint8_t> _pixels;
 
     GLuint             _fbo;
     GLuint             _textureID;
